smart_home/requirements.txt
dotenv
django~=5.1.1
djangorestframework~=3.15.2
psycopg2-binary
pillow
